+++ TaskManagement.Application/DTOs/RequestDTOs/Invitatoin/RevokeOrgInvitationAppDto.cs
@@ -1,5 +1,5 @@
 ﻿namespace TaskManagement.Application.DTOs.RequestDTOs.Invitatoin;
 public record RevokeOrgInvitationAppDto(
-    int OrgOwnerId,
-    int InvitationId
+    long OrgOwnerId,
+    long InvitationId
 );
