+++ TaskManagement.Application/DTOs/RequestDTOs/Organization/AddUserOrgAppDto.cs
@@ -1,5 +1,5 @@
 ﻿namespace TaskManagement.Application.DTOs.RequestDTOs.Organization;
 public record AddUserOrgAppDto(
-    int UserId,
-    int OrgId
+    long UserId,
+    long OrgId
 );
