+++ TaskManagement.Application/DTOs/RequestDTOs/Organization/ChangeActivityOrgAppDto.cs
@@ -1,7 +1,7 @@
 ﻿namespace TaskManagement.Application.DTOs.RequestDTOs.Organization;
 public record ChangeActivityOrgAppDto(
-    int OrgId,
-    int OwnerId,
+    long OrgId,
+    long OwnerId,
     string OwnerPassword,
     bool Activity
 );
