+++ TaskManagement.Application/DTOs/RequestDTOs/Organization/ChangeUserRoleOrgAppDto.cs
@@ -1,6 +1,6 @@
 ﻿namespace TaskManagement.Application.DTOs.RequestDTOs.Organization;
 public record ChangeUserRoleOrgAppDto(
-    int OrgOwnerId,
-    int OrgId,
-    int UserId
+    long OrgOwnerId,
+    long OrgId,
+    long UserId
 );
