+++ TaskManagement.Application/DTOs/RequestDTOs/Organization/CreateOrgAppDto.cs
@@ -3,6 +3,6 @@ public record CreateOrgAppDto(
     string OrgName,
     string SecondOrgName,
     string OrgDescription,
-    int OwnerId,
+    long OwnerId,
     byte MaxUser
 );
