+++ TaskManagement.Application/DTOs/RequestDTOs/Organization/RemoveUserOrgAppDto.cs
@@ -1,6 +1,6 @@
 ﻿namespace TaskManagement.Application.DTOs.RequestDTOs.Organization;
 public record RemoveUserOrgAppDto(
-    int OrgOwnerId,
-    int UserId,
-    int OrgId
+    long OrgOwnerId,
+    long UserId,
+    long OrgId
 );
