+++ TaskManagement.Application/DTOs/RequestDTOs/Organization/UpdateOrgAppDto.cs
@@ -1,7 +1,7 @@
 ﻿namespace TaskManagement.Application.DTOs.RequestDTOs.Organization;
 public record UpdateOrgAppDto(
-    int UserId,
-    int OrgId,
+    long UserId,
+    long OrgId,
     string OrgName,
     string SecondOrgName,
     string OrgDescription
