+++ TaskManagement.Application/DTOs/RequestDTOs/Task/UserTaskAppDto.cs
@@ -1,5 +1,5 @@
 ﻿namespace TaskManagement.Application.DTOs.RequestDTOs.Task;
 public record UserTaskAppDto(
-    int UserId,
-    int TaskId
+    long UserId,
+    long TaskId
 );
