+++ TaskManagement.Application/DTOs/RequestDTOs/TaskInfo/CreateTaskInfoAppDto.cs
@@ -1,5 +1,5 @@
 ﻿namespace TaskManagement.Application.DTOs.RequestDTOs.TaskInfo;
 public record CreateTaskInfoAppDto(
-    int TaskId,
-    int UserId
+    long TaskId,
+    long UserId
 );
