+++ TaskManagement.Application/Features/Invitation/Command/RevokeInvitation/RevokeInvitationCommand.cs
@@ -3,6 +3,6 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Invitation.Command.RevokeInvitation;
 public record RevokeInvitationCommand(
-    int OrgOwnerId,
-    int InvitationId
+    long OrgOwnerId,
+    long InvitationId
 ) : IRequest<GeneralResult>;
