+++ TaskManagement.Application/Features/Invitation/Query/GetAllOrgInvitationByOrgId/GetAllOrgInvitationByOrgIdQuery.cs
@@ -3,5 +3,5 @@ using TaskManagement.Application.DTOs.ResponseDTOs.Invitation;
 using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Invitation.Query.GetAllOrgInvitationByOrgId;
-public record GetAllOrgInvitationByOrgIdQuery(int OrgId)
+public record GetAllOrgInvitationByOrgIdQuery(long OrgId)
     : IRequest<GeneralResult<List<OrgInvitationDetailsDto>>>;
