+++ TaskManagement.Application/Features/Organization/Command/AddUserToOrg/AddUserToOrgCommand.cs
@@ -3,6 +3,6 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Organization.Command.AddUserToOrg;
 public record AddUserToOrgCommand(
-    int UserId,
-    int OrgId
+    long UserId,
+    long OrgId
 ) : IRequest<GeneralResult>;
