+++ TaskManagement.Application/Features/Organization/Command/LeaveUserFromOrg/LeaveUserFromOrgCommand.cs
@@ -3,6 +3,6 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Organization.Command.LeaveUserFromOrg;
 public record LeaveUserFromOrgCommand(
-    int UserId,
-    int OrgId
+    long UserId,
+    long OrgId
 ) : IRequest<GeneralResult>;
