+++ TaskManagement.Application/Features/Organization/Command/UpdateOrg/UpdateOrgCommand.cs
@@ -3,8 +3,8 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Organization.Command.UpdateOrg;
 public record UpdateOrgCommand(
-    int UserId,
-    int OrgId,
+    long UserId,
+    long OrgId,
     string OrgName,
     string SecondOrgName,
     string OrgDescription
