+++ TaskManagement.Application/Features/Project/Command/RemoveUserFromProject/RemoveUserFromProjectCommand.cs
@@ -3,7 +3,7 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Project.Command.RemoveUserFromProject;
 public record RemoveUserFromProjectCommand(
-    int UserId,
-    int ProjId,
-    int OwnerId
+    long UserId,
+    long ProjId,
+    long OwnerId
 ) : IRequest<GeneralResult>;
