+++ TaskManagement.Application/Features/Task/Command/AssignUserToTask/AssignUserToTaskCommand.cs
@@ -3,8 +3,8 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Task.Command.AssignUserToTask;
 public record AssignUserToTaskCommand(
-    int OwnerId,
-    int UserId,
-    int TaskId,
-    int ProjId
+    long OwnerId,
+    long UserId,
+    long TaskId,
+    long ProjId
 ) : IRequest<GeneralResult>;
