+++ TaskManagement.Application/Features/Task/Command/ChangeTaskType/ChangeTaskTypeCommand.cs
@@ -3,6 +3,6 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Task.Command.ChangeTaskType;
 public record ChangeTaskTypeCommand(
-    int UserId,
-    int TaskId
+    long UserId,
+    long TaskId
 ) : IRequest<GeneralResult>;
