+++ TaskManagement.Application/Features/Task/Command/EndTask/EndTaskCommand.cs
@@ -3,6 +3,6 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Task.Command.EndTask;
 public record EndTaskCommand(
-    int UserId,
-    int TaskId
+    long UserId,
+    long TaskId
 ) : IRequest<GeneralResult>;
