+++ TaskManagement.Application/Features/Task/Command/RemoveUserFromTask/RemoveUserFromTaskCommand.cs
@@ -3,8 +3,8 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.Task.Command.RemoveUserFromTask;
 public record RemoveUserFromTaskCommand(
-    int OwnerId,
-    int UserId,
-    int TaskId,
-    int ProjId
+    long OwnerId,
+    long UserId,
+    long TaskId,
+    long ProjId
 ) : IRequest<GeneralResult>;
