+++ TaskManagement.Application/Features/User/Command/DecreaseUserPoints/DecreaseUserPointsCommand.cs
@@ -3,5 +3,5 @@ using TaskManagement.Common.Classes;
 
 namespace TaskManagement.Application.Features.User.Command.DecreaseUserPoints;
 
-public record DecreaseUserPointsCommand(int UserId)
+public record DecreaseUserPointsCommand(long UserId)
     : IRequest<GeneralResult>;
